+++ CHANGELOG.md
@@ -18,6 +18,7 @@ For a full diff see [`1.0.3...main`][1.0.3...main].
 - Extracted `Format\Format::create()` as named constructor and reduced visibility of `__construct`  to `private` ([#608]), by [@localheinz]
 - Stopped composing `Format\Format` into `Json` ([#616]), by [@localheinz]
 - Renamed `FinalNewLineNormalizer` to `WithFinalNewLineNormalizer` ([#618]), by [@localheinz]
+- Renamed `NoFinalNewLineNormalizer` to `WithoutFinalNewLineNormalizer` ([#619]), by [@localheinz]
 
 ### Fixed
 
@@ -442,6 +443,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#608]: https://github.com/ergebnis/json-normalizer/pull/608
 [#616]: https://github.com/ergebnis/json-normalizer/pull/616
 [#618]: https://github.com/ergebnis/json-normalizer/pull/618
+[#619]: https://github.com/ergebnis/json-normalizer/pull/619
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@dependabot]: https://github.com/dependabot